This repo stores the official Oracle Linux base images in the form of tarballs containing a rootfs in two architecture-specific branches. These images are automatically consumed by the Docker Hub and are not intended for direct use by end-users.
Due to the fact that only the latest tarball for each tag is stored in the dist-amd64 and dist-arm64v8 branches, it is not possible to use Git's history to determine what has changed and when. Thus, we maintain a CHANGELOG in the master branch that documents the changes made to various images by date.

Signed-off-by: Avi Miller <avi.miller@oracle.com>